backend/requirements.txt

tensorflow==2.13.0
tensorflow-hub==0.13.0
keras==2.13.1
numpy
pandas
matplotlib
seaborn
scikit-learn
flask
werkzeug
opencv-python
Pillow
flask-cors
python-dotenv
speechrecognition==3.10.0
requests
Flask-JWT-Extended
bcrypt
itsdangerous
imutils
flask-pymongo
